requirements.txt
streamlit
pandas
fpdf==1.7.2
openpyxl
office365-rest-python-client